+++ CHANGES.rst
@@ -6,6 +6,9 @@ Alterações
 
   * Revisado tamanho do footer dos tiles. (closes `#88`_).
     [rodfersou][rennanrodrigues]
+  * Adicionada opção para variação de título nos tiles basic e
+    basic news article (nitf) (closes `#86`_).
+    [rodfersou]
   * Adicionado suporte ao scale de imagem original (closes `#82`_).
     [rodfersou]
 
@@ -147,4 +150,5 @@ Alterações
 .. _`#72`: https://github.com/plonegovbr/brasil.gov.tiles/issues/72
 .. _`#74`: https://github.com/plonegovbr/brasil.gov.tiles/issues/74
 .. _`#82`: https://github.com/plonegovbr/brasil.gov.tiles/issues/82
+.. _`#86`: https://github.com/plonegovbr/brasil.gov.tiles/issues/86
 .. _`#88`: https://github.com/plonegovbr/brasil.gov.tiles/issues/88